add opp to choose role for new users

+++ src/main/java/ru/kata/spring/boot_security/demo/model/Role.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column(name = "role")
     private String name;
@@ -26,12 +26,12 @@ public class Role implements GrantedAuthority {
         this.name = name;
     }
 
-    public Role(Integer id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
